:wrench: #46 createBusinessInfo userId 타입이슈 개선

+++ src/main/kotlin/kr/co/marketbill/marketbillcoreserver/service/UserService.kt
@@ -168,7 +168,7 @@ class UserService {
 
     fun createBusinessInfo(input: CreateBusinessInfoInput): BusinessInfo {
         val businessInfo = BusinessInfo(
-            user = entityManager.getReference(User::class.java, input.userId),
+            user = entityManager.getReference(User::class.java, input.userId.toLong()),
             companyName = input.companyName,
             companyPhoneNo = input.companyPhoneNo,
             businessMainCategory = input.businessMainCategory,
